backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
Pillow==10.3.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
python-dotenv==1.0.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
django-simpleui==2026.1.13
pymysql==1.1.0